Testing moving file
